# Optimisation in excprob function